+++ src/index.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="stylesheet" href="app.css">
-  <script defer src="app.js"></script>
+  <script type="module" src="./app.js"></script>
   <title>werewolf client</title>
 </head>
 <body>
